This commit was generated by cvs2svn to compensate for changes in r4023, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@4024 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/6510core.c
@@ -1430,7 +1430,7 @@
 #  if !defined WORDS_BIGENDIAN && defined ALLOW_UNALIGNED_ACCESS
 
 #    define opcode_t DWORD
-#    define FETCH_OPCODE(o) ((o) = ((reg_pc < bank_limit)                   \
+#    define FETCH_OPCODE(o) ((o) = ((((int)reg_pc) < bank_limit)           \
                                     ? (*((DWORD *)(bank_base + reg_pc))    \
                                        & 0xffffff)                         \
                                     : (LOAD(reg_pc)                        \
